requirements.txt
streamlit==1.32.2
pandas==2.0.3
numpy==1.26.4
pdfplumber==0.10.3
python-docx==0.8.11
regex==2023.10.3
pillow==10.4.0
